test(codex): isolate runtime teardown coverage
